requirements.txt
# HC-AI MCP Server - Dependencies

# MCP Server
mcp>=1.0.0

# Web framework (used by MCP for HTTP transport)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0

# Database (PostgreSQL + pgvector)
sqlalchemy>=2.0.0
asyncpg>=0.29.0
langchain-postgres>=0.0.1

# Embeddings and LLM
langchain-core>=0.1.0
langchain-ollama>=0.1.0
langchain-aws>=0.1.0  # For Bedrock support
langchain-openai>=0.1.0  # For OpenAI support
langchain-anthropic>=0.1.0  # For Anthropic support
langchain-text-splitters>=0.0.1

# Agent workflow
langgraph>=0.2.0

# Reranking
sentence-transformers>=2.2.0
torch>=2.0.0

# Text processing
nltk>=3.8
numpy>=1.24.0
scikit-learn>=1.3.0  # For cosine similarity

# HTTP client
requests>=2.31.0
httpx>=0.27.0

# Safety & Validation
guardrails-ai>=0.4.0
pydeid

# AWS (optional, for Bedrock and DynamoDB)
boto3>=1.34.0

# Token counting
tiktoken>=0.5.0

# Development / Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.23.0
